Extract `QueryKind` into the types
